+++ packages/decap-cms-core/src/types/redux.ts
@@ -149,6 +149,8 @@ export interface CmsFieldList {
   default?: unknown;
 
   allow_add?: boolean;
+  allow_remove?: boolean;
+  allow_reorder?: boolean;
   collapsed?: boolean;
   summary?: string;
   minimize_collapsed?: boolean;
@@ -415,7 +417,11 @@ export interface CmsConfig {
   locale?: string;
   site_url?: string;
   display_url?: string;
-  logo_url?: string;
+  logo_url?: string; // Deprecated, replaced by `logo.src`
+  logo?: {
+    src: string;
+    show_in_header?: boolean;
+  };
   show_preview_links?: boolean;
   media_folder?: string;
   public_folder?: string;